requirements.txt
pydantic==1.7.3
fastapi==0.62.0
PyJWT==1.7.1
bcrypt==3.2.0
